task: migrate business logic from cogs to services
